Create Peripheral Controller
Problem statement: A peripheral controller is triggered by a signal start from a host processor. On sensing the start, the controller is expected to perform the following actions
Assert an acknowledge signal ack for one clock cycle
Assert a data_req signal to a peripheral requesting data
Wait for a data_avl (indicating availability of data from the peripheral) handshake from the peripheral.
On sensing the data_avl signal send a req_grant signal to the host. If the data_avl is not asserted for 100 clocks, then raise an exception through the interrupt line to the host, intr
Draw an RTL of the system explaining the blocks 
Model the controller state machine in Verilog